sam normalized and RGB float 16

+++ src/vollseg/utils.py
@@ -4541,7 +4541,7 @@ def VollSamZ(
         channel_image = channel_image * 255
 
         instance_labels_currentz = mask_generator.generate(
-            channel_image.astype(np.float32)
+            channel_image.astype(np.float16)
         )
         instance_labels_currentz = remove_small_objects(
             instance_labels_currentz.astype("uint16"), min_size=min_size
